+++ app/linux/flutter/generated_plugin_registrant.cc
@@ -10,6 +10,7 @@
 #include <dynamic_color/dynamic_color_plugin.h>
 #include <file_selector_linux/file_selector_plugin.h>
 #include <gtk/gtk_plugin.h>
+#include <open_dir_linux/open_dir_linux_plugin.h>
 #include <pasteboard/pasteboard_plugin.h>
 #include <screen_retriever/screen_retriever_plugin.h>
 #include <tray_manager/tray_manager_plugin.h>
@@ -31,6 +32,9 @@ void fl_register_plugins(FlPluginRegistry* registry) {
   g_autoptr(FlPluginRegistrar) gtk_registrar =
       fl_plugin_registry_get_registrar_for_plugin(registry, "GtkPlugin");
   gtk_plugin_register_with_registrar(gtk_registrar);
+  g_autoptr(FlPluginRegistrar) open_dir_linux_registrar =
+      fl_plugin_registry_get_registrar_for_plugin(registry, "OpenDirLinuxPlugin");
+  open_dir_linux_plugin_register_with_registrar(open_dir_linux_registrar);
   g_autoptr(FlPluginRegistrar) pasteboard_registrar =
       fl_plugin_registry_get_registrar_for_plugin(registry, "PasteboardPlugin");
   pasteboard_plugin_register_with_registrar(pasteboard_registrar);