+++ app/linux/flutter/generated_plugins.cmake
@@ -7,6 +7,7 @@ list(APPEND FLUTTER_PLUGIN_LIST
   dynamic_color
   file_selector_linux
   gtk
+  open_dir_linux
   pasteboard
   screen_retriever
   tray_manager