+++ app/assets/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.16.1 (unreleased)
 
+- feat(desktop): highlight file when tapping "Show in folder" (@Tienisto)
 - fix(android): properly close app on back gesture (@Tienisto)
 
 ## 1.16.0 (2024-11-03)